Extract some code from int.h to intttypes.h to avoid circular dependencies (#39)
